moved all library code into a folder for neatness. Added readme and some images
