+++ .gitignore
@@ -37,4 +37,5 @@ test_XXZ.py
 train_magic/
 training_tests.py
 previous_project_code/*
+images/*
